components: organize css files into single directory
