+++ src/components/css/BaseLayout.css
@@ -1,0 +1,13 @@
+@import url("https://cdnjs.cloudflare.com/ajax/libs/meyer-reset/2.0/reset.min.css");
+@import url("https://fonts.googleapis.com/css?family=Raleway:500,600,400,700|Epilogue:700|Lato:700");
+
+* {
+  box-sizing: border-box;
+}
+
+.sidebar-handle {
+  height: 100vh;
+  background-color: #1b1c1d;
+  padding: 1em;
+  width: 5em;
+}
